Change the way the CO2 inversion smoother works
Instead of a time_window_length, take cycle_num as an input from the
namelist to determine which observations to assimilate and CO2 state
variable to update.

The reason for this change is that the CO2 state is not updated when
assimilating observations. If this is caused by a bug that can be fixed,
then this commit can be reverted.